Mega-pint update
Tons of refactoring. Added video playback, and other goodies.

+++ src/util/Implode.c
@@ -353,7 +353,7 @@ static unsigned long Expand(TDcmpStruct *pWork) {
         }
     }
 
-    copyBytes = pWork->outputPos - 0x1000;
+    copyBytes = (unsigned int)pWork->outputPos - 0x1000;
     pWork->write_buf((char *)&pWork->out_buff[0x1000], &copyBytes, pWork->param);
     return dwResult;
 }
